+++ include/libpldm/entity.h
@@ -1,6 +1,5 @@
 /* SPDX-License-Identifier: Apache-2.0 OR GPL-2.0-or-later */
-#ifndef ENTITY_H
-#define ENTITY_H
+#pragma once
 
 #include <libpldm/api.h>
 
@@ -202,5 +201,3 @@ enum pldm_entity_id_codes {
 #ifdef __cplusplus
 }
 #endif
-
-#endif /* ENTITY_H */