+++ include/libpldm/oem/meta/file_io.h
@@ -1,6 +1,5 @@
 /* SPDX-License-Identifier: Apache-2.0 OR GPL-2.0-or-later */
-#ifndef LIBPLDM_OEM_META_FILE_IO_H
-#define LIBPLDM_OEM_META_FILE_IO_H
+#pragma once
 
 #include <libpldm/compiler.h>
 
@@ -164,5 +163,3 @@ int encode_oem_meta_file_io_read_resp(
 #ifdef __cplusplus
 }
 #endif
-
-#endif /*LIBPLDM_OEM_META_FILE_IO_H*/